Use MUnit in ClientParserSuite
